+++ src/main/java/com/ureca/yoajungserver/review/dto/response/ReviewCreateResponse.java
@@ -1,4 +1,4 @@
-package com.ureca.yoajungserver.review.dto;
+package com.ureca.yoajungserver.review.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,7 +9,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class ReviewUpdateRequest {
-    private String content;
-    private int star;
+public class ReviewCreateResponse {
+    Long reviewId;
 }
