+++ src/main/java/com/ureca/yoajungserver/review/repository/custom/ReviewRepositoryCustom.java
@@ -1,6 +1,6 @@
 package com.ureca.yoajungserver.review.repository.custom;
 
-import com.ureca.yoajungserver.review.dto.ReviewListResponse;
+import com.ureca.yoajungserver.review.dto.response.ReviewListResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
