+++ src/main/java/com/ureca/yoajungserver/review/service/ReviewService.java
@@ -1,12 +1,14 @@
 package com.ureca.yoajungserver.review.service;
 
-import com.ureca.yoajungserver.review.dto.*;
+import com.ureca.yoajungserver.review.dto.request.ReviewCreateRequest;
+import com.ureca.yoajungserver.review.dto.request.ReviewUpdateRequest;
+import com.ureca.yoajungserver.review.dto.response.*;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
 public interface ReviewService {
     Page<ReviewListResponse> reviewList(Long planId, Pageable pageable);
-    ReviewCreateResponse insertReview(ReviewCreateRequest request);
+    ReviewCreateResponse insertReview(Long planId, ReviewCreateRequest request);
     ReviewUpdateResponse updateReview(Long reviewId, ReviewUpdateRequest request);
     ReviewDeleteResponse deleteReview(Long reviewId);
